refactor: rename ApprovalService to PublicationManagerService and removed classes in the ApprovalsService module as this does not need to be tracked.
