--- unicorn_web/PropertyFunctions/src/main/java/property/dao/Property.java
@@ -1,1 +1,0 @@
-
