+++ unicorn_web/PublicationManagerService/src/main/java/schema/unicorn_properties/publicationevaluationcompleted/AWSEvent.java
